moved profile wheel to own component, created main tag in MyProfile, started searchbar
