+++ GnuOne/ClientApp/src/components/NewComponents/search.css
@@ -1,0 +1,31 @@
+
+
+.search {
+    display: flex;
+    flex-direction: row;
+    align-items: flex-end;
+    padding-left: 5em;
+}
+.search input {
+    border: 1px rgb(255, 255, 255) solid;
+    width: 7em;
+    align-self: center;
+    padding: 0.5em;
+    margin: 0.5em 1.5em;
+    border-radius: 9px;
+    box-shadow: -3px 1px 3px rgb(226 226 226);
+}
+  
+.search input:focus {
+    outline:none;
+}
+
+.search button {
+    margin: 0;
+    font-style: bold;
+    color: rgba(124, 124, 124, 0.568);
+    text-decoration: none;
+    border: 1px gray solid;
+}
+
+
