+++ GnuOne/appsettings.json
@@ -8,6 +8,6 @@
   },
   "AllowedHosts": "*",
   "ConnectionStrings": {
-    "Defaultconnection": "server=localhost;user id=;password=;database=gnu;"
+    "Defaultconnection": "server=localhost;user id=root;password=johanna;database=gnu;"
   }
 }
